修改年份是生肖数字的归属,获取当前时间的数据

+++ hk6domain/src/main/java/com/joye/hk6domain/constants/Hk6EnumHelp.java
@@ -186,7 +186,7 @@ public class Hk6EnumHelp {
         return HeadTwos.HeadEven0;
     }
     
-    public static  ChineseZodiac getChineseZodiac(int n ,long time){
+    public static  ChineseZodiac getChineseZodiac(int n ,int time){
         for (ChineseZodiac mChineseZodiac:ChineseZodiac.values()){
             for (int item:mChineseZodiac.getNumbers(time)){
                 if(item==n){
